scale Testing

+++ Letter.h
@@ -13,13 +13,13 @@ class Letter
 
 		bool loadFromRenderedText(char letter, SDL_Color letterColor, SDL_Renderer* renderer);
 
-		void render(int x, int y,SDL_Renderer* renderer, SDL_Rect* clip=NULL, double angle = 0.0, SDL_Point* center = NULL, SDL_RendererFlip flip = SDL_FLIP_NONE);	
+		void render(double x, double y,SDL_Renderer* renderer, SDL_Rect* clip=NULL, double angle = 0.0, SDL_Point* center = NULL, SDL_RendererFlip flip = SDL_FLIP_NONE);	
 	
 		void fade();
 		
 		void setX(double x);
 
-		void setY(int Y);
+		void setY(double Y);
 
 		Uint8 getAlpha();	
 
@@ -31,7 +31,7 @@ class Letter
 		
 		double getX();
 
-		int getY();
+		double getY();
 
     void setDynamic(int value); 
     
@@ -50,17 +50,17 @@ class Letter
 		SDL_Texture* shadowTexture = NULL;
 		TTF_Font* font = NULL;
 		char letter = '\0';
-		int width = 0;
-		int height= 0;	
+		double width = 0;
+		double height= 0;	
 		double xPos= 0;
-		int yPos= 0;
+		double yPos= 0;
 		double relX=0;
 		int randGroundValue=0;
-		int relY=0;
-	  int prevRelY=0;	
-    int rate=0;
+		double relY=0;
+	  double prevRelY=0;	
+    double rate=0;
     int randXChange=0;
-    int velocity=0;
+    double velocity=0;
 		int angle=0;
 		Uint32 lifetime = 0;
 		bool onGround = false;
